ref(issue-workflow): Move build_activity_notification_data to template layer
Move build_activity_notification_data and extract_notification_models_by_activity
from the action registry into the template layer (activity/base.py) where they
belong. Make workflow_id optional so the builder can be used without a workflow
context. Remove notification_uuid from ActivityNotificationData. Update all
imports and tests accordingly.
