+++ src/sentry/notifications/platform/templates/activity/__init__.py
@@ -1,11 +1,4 @@
 from .assigned import AssignedActivityTemplate
-from .base import (
-    ACTIVITY_TYPE_TO_SOURCE,
-    ActivityNotificationData,
-    AssignedNotificationData,
-    SetResolvedInCommitNotificationData,
-    SetResolvedInReleaseNotificationData,
-)
 from .note import NoteActivityTemplate
 from .seer.coding_completed import SeerCodingCompletedActivityTemplate
 from .seer.coding_started import SeerCodingStartedActivityTemplate
@@ -27,11 +20,6 @@ from .status_change.set_unresolved import SetUnresolvedActivityTemplate
 from .unassigned import UnassignedActivityTemplate
 
 __all__ = (
-    "ACTIVITY_TYPE_TO_SOURCE",
-    "ActivityNotificationData",
-    "SetResolvedInCommitNotificationData",
-    "SetResolvedInReleaseNotificationData",
-    "AssignedNotificationData",
     "SeerRcaStartedActivityTemplate",
     "SeerRcaCompletedActivityTemplate",
     "SeerSolutionStartedActivityTemplate",
