+++ src/sentry/notifications/platform/templates/activity/base.py
@@ -1,5 +1,16 @@
 from typing import Any
+from urllib.parse import urlencode
 
+from sentry.integrations.messaging.message_builder import (
+    build_attachment_text,
+    build_attachment_title,
+)
+from sentry.models.activity import Activity
+from sentry.models.commit import Commit
+from sentry.models.group import Group
+from sentry.models.organization import Organization
+from sentry.models.project import Project
+from sentry.notifications.notifications.activity.assigned import get_assignee_str
 from sentry.notifications.platform.types import (
     CodeSection,
     CodeTextBlock,
@@ -11,7 +22,10 @@ from sentry.notifications.platform.types import (
     ParagraphSection,
     PlainTextBlock,
 )
-from sentry.types.activity import ActivityType
+from sentry.types.activity import SEER_ACTIVITY_TYPES, ActivityType
+from sentry.users.services.user.service import user_service
+from sentry.utils.http import absolute_uri
+from sentry.workflow_engine.models import Workflow
 
 ACTIVITY_TYPE_TO_SOURCE: dict[int, NotificationSource] = {
     ActivityType.SEER_RCA_STARTED.value: NotificationSource.ACTIVITY_SEER_RCA_STARTED,
@@ -45,7 +59,6 @@ FOOTER_DELIMITER = " · "
 
 class ActivityNotificationData(NotificationData):
     source: NotificationSource
-    notification_uuid: str
     activity_type: int
     activity_data: dict[str, Any] | None = None
     # The name of the user who is associated with an activity (Activity.user_id)
@@ -69,7 +82,6 @@ def create_activity_notification_example(
     activity_data: dict[str, Any] | None = None,
 ) -> ActivityNotificationData:
     return ActivityNotificationData(
-        notification_uuid="1234567890",
         activity_user_name="Jane Doe",
         issue_short_id="JAVASCRIPT-1",
         issue_url=EXAMPLE_ISSUE_URL,
@@ -129,3 +141,113 @@ def get_issue_description(data: ActivityNotificationData) -> list[NotificationSe
     if data.issue_description:
         sections.append(CodeSection(blocks=[PlainTextBlock(text=data.issue_description)]))
     return sections
+
+
+def extract_notification_models_by_activity(
+    activity: Activity,
+) -> tuple[Group, Project, Organization]:
+    try:
+        group = Group.objects.get_from_cache(id=activity.group_id)
+    except Group.DoesNotExist:
+        raise ValueError(f"Group not found: {activity.group_id}")
+    try:
+        project = Project.objects.get_from_cache(id=activity.project_id)
+    except Project.DoesNotExist:
+        raise ValueError(f"Project not found: {activity.project_id}")
+    try:
+        organization = Organization.objects.get_from_cache(id=project.organization_id)
+    except Organization.DoesNotExist:
+        raise ValueError(f"Organization not found: {project.organization_id}")
+
+    return group, project, organization
+
+
+def build_activity_notification_data(
+    activity: Activity, *, workflow_id: int | None = None
+) -> ActivityNotificationData:
+    source = ACTIVITY_TYPE_TO_SOURCE.get(activity.type)
+    if source is None:
+        raise ValueError(f"No notification source for activity type: {activity.type}")
+
+    group, project, organization = extract_notification_models_by_activity(activity)
+
+    workflow: Workflow | None = None
+    if workflow_id:
+        try:
+            workflow = Workflow.objects.get(id=workflow_id, organization_id=organization.id)
+        except Workflow.DoesNotExist:
+            raise ValueError(f"Workflow not found: {workflow_id}")
+
+    issue_url_params: dict[str, str] = {}
+    if ActivityType(activity.type) in SEER_ACTIVITY_TYPES:
+        issue_url_params.update({"seerDrawer": "true"})
+
+    action_data = dict(
+        source=source,
+        activity_type=activity.type,
+        issue_short_id=group.qualified_short_id,
+        issue_url=absolute_uri(group.get_absolute_url(params=issue_url_params)),
+        issue_title=build_attachment_title(group) or "",
+        issue_culprit=group.culprit,
+        issue_description=build_attachment_text(group),
+        project_slug=project.slug,
+        project_url=organization.absolute_url(
+            f"organizations/{organization.slug}/issues/",
+            query=urlencode({"project": project.id}),
+        ),
+        activity_data=activity.data,
+        activity_user_name=None,
+    )
+
+    if workflow:
+        action_data.update(
+            {
+                "alert_name": workflow.name,
+                "alert_url": organization.absolute_url(
+                    f"organizations/{organization.slug}/monitors/alerts/{workflow_id}/"
+                ),
+            }
+        )
+
+    if activity.user_id:
+        user = user_service.get_user(user_id=activity.user_id)
+        if user:
+            action_data["activity_user_name"] = user.get_display_name()
+
+    match activity.type:
+        case ActivityType.SET_RESOLVED_IN_COMMIT.value:
+            commit_sha = None
+            commit_message = None
+            if activity.data and "commit" in activity.data:
+                try:
+                    commit = Commit.objects.get(id=activity.data["commit"])
+                    commit_sha = commit.short_id
+                    commit_message = commit.message
+                except Commit.DoesNotExist:
+                    pass
+            return SetResolvedInCommitNotificationData(
+                **action_data, commit_sha=commit_sha, commit_message=commit_message
+            )
+        case ActivityType.SET_RESOLVED_IN_RELEASE.value:
+            release_url = None
+            # If version is missing, None or "" -> it was resolved in an upcoming release
+            if activity.data and activity.data.get("version"):
+                raw_version = activity.data["version"]
+                release_url = organization.absolute_url(
+                    f"organizations/{organization.slug}/releases/{raw_version}/",
+                    query=urlencode({"project": project.id}),
+                )
+            return SetResolvedInReleaseNotificationData(**action_data, release_url=release_url)
+
+        case ActivityType.ASSIGNED.value:
+            assignee_label = get_assignee_str(activity=activity, organization=organization)
+            assignee_email = activity.data.get("assigneeEmail")
+            assignee_url = None
+            # TODO(Leander): If a team is assigned, maybe link to the team page?
+            if assignee_email:
+                assignee_url = f"mailto:{assignee_email}"
+            return AssignedNotificationData(
+                **action_data, assignee_label=assignee_label, assignee_url=assignee_url
+            )
+        case _:
+            return ActivityNotificationData(**action_data)
